+++ domain/dropbox/api/dbx_error/error.go
@@ -40,6 +40,7 @@ type Errors interface {
 	Endpoint() ErrorEndpoint
 	To() ErrorWrite
 	BadPath() ErrorSharePath
+	Member() ErrorMember
 
 	// too_many_write_operations
 	IsTooManyWriteOperations() bool
@@ -53,6 +54,8 @@ type Errors interface {
 	IsDuplicateUser() bool
 	// member not in group
 	IsMemberNotInGroup() bool
+	// id not found
+	IsIdNotFound() bool
 
 	Summary() string
 }
@@ -110,8 +113,17 @@ type ErrorSharePath interface {
 	IsAlreadyShared() bool
 }
 
+// 409: ErrorMember
+type ErrorMember interface {
+	IsNotAMember() bool
+}
+
 type errorsImpl struct {
 	de DropboxError
+}
+
+func (z errorsImpl) IsIdNotFound() bool {
+	return z.de.HasPrefix("id_not_found")
 }
 
 func (z errorsImpl) IsMemberNotInGroup() bool {
@@ -164,4 +176,8 @@ func (z errorsImpl) Endpoint() ErrorEndpoint {
 
 func (z errorsImpl) To() ErrorWrite {
 	return NewErrorWrite("to", z.de)
+}
+
+func (z errorsImpl) Member() ErrorMember {
+	return NewErrorMember(z.de)
 }