+++ essentials/strings/es_mailaddr/mailaddr.go
@@ -1,5 +1,7 @@
 package es_mailaddr
 
+import "regexp"
+
 func EscapeSpecial(email string, alt string) string {
 	// RFC 5322, atext (excl ALPHA/DIGIT)
 	specials := []rune("!#$%&'*+-/=?^_`{|}~")
@@ -21,4 +23,12 @@ func EscapeSpecial(email string, alt string) string {
 		}
 	}
 	return string(e)
+}
+
+var (
+	emailPattern = regexp.MustCompile("^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$")
+)
+
+func IsEmailAddr(email string) bool {
+	return emailPattern.MatchString(email)
 }