+++ recipe/member/update/visibility_record.go
@@ -1,0 +1,5 @@
+package update
+
+type VisibilityRecord struct {
+	Email string
+}