+++ recipe/teamfolder/member/list.go
@@ -7,7 +7,7 @@ import (
 	"github.com/watermint/toolbox/domain/dropbox/service/sv_profile"
 	"github.com/watermint/toolbox/domain/dropbox/usecase/uc_folder_member"
 	"github.com/watermint/toolbox/domain/dropbox/usecase/uc_team_content"
-	"github.com/watermint/toolbox/domain/dropbox/usecase/uc_teamfolder"
+	"github.com/watermint/toolbox/domain/dropbox/usecase/uc_teamfolder_scanner"
 	"github.com/watermint/toolbox/essentials/kvs/kv_kvs"
 	"github.com/watermint/toolbox/essentials/kvs/kv_storage"
 	"github.com/watermint/toolbox/essentials/model/mo_filter"
@@ -62,9 +62,9 @@ func (z *List) Preset() {
 		z.memberTypeInternal,
 		z.memberTypeExternal,
 	)
-	z.ScanTimeout.SetOptions(string(uc_teamfolder.ScanTimeoutShort),
-		string(uc_teamfolder.ScanTimeoutShort),
-		string(uc_teamfolder.ScanTimeoutLong),
+	z.ScanTimeout.SetOptions(string(uc_teamfolder_scanner.ScanTimeoutShort),
+		string(uc_teamfolder_scanner.ScanTimeoutShort),
+		string(uc_teamfolder_scanner.ScanTimeoutLong),
 	)
 }
 
@@ -81,7 +81,7 @@ func (z *List) Exec(c app_control.Control) error {
 		return err
 	}
 
-	teamFolderScanner := uc_teamfolder.New(c, z.Peer.Context(), uc_teamfolder.ScanTimeoutMode(z.ScanTimeout.Value()))
+	teamFolderScanner := uc_teamfolder_scanner.New(c, z.Peer.Context(), uc_teamfolder_scanner.ScanTimeoutMode(z.ScanTimeout.Value()))
 	teamFolders, err := teamFolderScanner.Scan(z.Folder)
 	if err != nil {
 		return err