+++ recipe/teamfolder/policy/list.go
@@ -3,7 +3,7 @@ package policy
 import (
 	"github.com/watermint/toolbox/domain/dropbox/api/dbx_conn"
 	"github.com/watermint/toolbox/domain/dropbox/usecase/uc_team_content"
-	"github.com/watermint/toolbox/domain/dropbox/usecase/uc_teamfolder"
+	"github.com/watermint/toolbox/domain/dropbox/usecase/uc_teamfolder_scanner"
 	"github.com/watermint/toolbox/essentials/log/esl"
 	"github.com/watermint/toolbox/essentials/model/mo_filter"
 	"github.com/watermint/toolbox/essentials/model/mo_string"
@@ -34,14 +34,14 @@ func (z *List) Preset() {
 		mo_filter.NewNamePrefixFilter(),
 		mo_filter.NewNameSuffixFilter(),
 	)
-	z.ScanTimeout.SetOptions(string(uc_teamfolder.ScanTimeoutShort),
-		string(uc_teamfolder.ScanTimeoutShort),
-		string(uc_teamfolder.ScanTimeoutLong),
+	z.ScanTimeout.SetOptions(string(uc_teamfolder_scanner.ScanTimeoutShort),
+		string(uc_teamfolder_scanner.ScanTimeoutShort),
+		string(uc_teamfolder_scanner.ScanTimeoutLong),
 	)
 }
 
 func (z *List) Exec(c app_control.Control) error {
-	teamFolderScanner := uc_teamfolder.New(c, z.Peer.Context(), uc_teamfolder.ScanTimeoutMode(z.ScanTimeout.Value()))
+	teamFolderScanner := uc_teamfolder_scanner.New(c, z.Peer.Context(), uc_teamfolder_scanner.ScanTimeoutMode(z.ScanTimeout.Value()))
 	teamFolders, err := teamFolderScanner.Scan(z.Folder)
 	if err != nil {
 		return err